pyproject.toml
[tool.poetry]
name = "buildmetadata"
version = "0.1.0"
description = ""
authors = ["Max Schneider <ms.matix@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tk = "^0.1.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
